+++ src/puppeteer.js
@@ -242,7 +242,8 @@ export const launchPuppeteer = async (options = {}) => {
         optsCopy.args.push(`--user-agent=${userAgent}`);
     }
 
-    if (optsCopy.pipe !== false) optsCopy.pipe = true;
+    // TODO Monitor https://github.com/GoogleChrome/puppeteer/issues/4374
+    // if (optsCopy.pipe !== false) optsCopy.pipe = true;
 
     let browser;
     if (optsCopy.proxyUrl) {